Rename derive_variables -> derived_variables
